fix(workset): enforce canonical anchor invariants
